Add support for backup to URL instead of blob storage
8/6/2016... for version V002.

Update scripts to support backup to an URL like \\Servername\Sharename
or \\Servername\Sharename\Whatever

done: Update sp_ScheduleBackups to accept an @urlPath parameter.  This
will be mutually exclusive with @StorageContainerURL.
done: Update sp_ScheduleBackups to check if an @urlPath param has been
supplied and that it looks like \\someServer\someShare
done: Update sp_ScheduleBackups to accept optional
@intFullBackupExpireDays, @intDiffBackupExpireDays and
@intLogBackupExpireDays parameters.  If supplied these will be used with
Ola's scripts to expire the backups.
done: Update sp_ScheduleBackups to create jobs as backup to UNC paths
instead of URL
done: Update usp_RestoreDB to have optional parameter for
BlobCredential.  If possible have it check to see if any of the files to
be restored are from an URL path and complain about not having the
parameter if necessary.
done: Update usp_AutoMirror to have optional parameter for
BlobCredential.  If possible have it check to see if any of the files to
be restored are from an URL path and complain about not having the
parameter if necessary.
done: Make changes so that all of the above param changes work :)
done: Update documentation